account game dto list set current tipe to balance games
-double
+int

+++ SmartFunRevamp/lib/core/domain/entities/card_details/account_game_dto_list.dart
@@ -6,7 +6,7 @@ class AccountGameDTOList {
   final DateTime? fromDate;
   final DateTime? expiryDate;
   final double quantity;
-  final double balanceGames;
+  final int balanceGames;
   final int gameId;
   final int gameProfileId;
